Update autoload.php
It took me a while to work it out, not being that familiar with code.  

There have been a lot of users trying to get a resolution to the "weather station plugin PHP Deprecated: file_exists(): Passing null to parameter #1 ($filename) of type string is deprecated in /home/burramot/public_html/wp-content/plugins/live-weather-station/autoload.php on line 246" error.  This error was being by the conditional statement (3 lines) at the end of the code.

With later versions of PHP this was causing excessive information error logging alerting the developer to the depracated code, which in turn was dramatically affecting overall performance of the website.

Declaring the variable with a null value resolves this issue.

Please correct code as suggested update and release an update ASAP.